#9 메인디테일 및 화면 전반적 구현

+++ data/dto/src/main/java/RequestModel/SignUpRequestModel.kt
@@ -3,8 +3,8 @@ package RequestModel
 data class SignUpRequestModel(
     val oauthToken: String,
     val oauthProvider: String = "kakao",
-    val mainRegionId: Long,
-    val interestedRegionIds: List<Long>,
+    val mainRegionId: Int,
+    val interestedRegionIds: List<Int>,
     val nickname: String,
     val agreements: List<String>
 )
